WebScraper/requirements.txt
# These are the dependencies needed to run the web scraper
selenium==4.25.0
wakepy==0.10.1
webdriver-manager==4.0.2
python-dotenv==1.0.1
pymongo==4.11.1